Moves format_uptime to utils module.
